Rename terminal->shell
