+++ apps/sh/utils.c
