+++ apps/sh/values.c
@@ -21,7 +21,7 @@
  **   Amanuel Bogale <amanuel2> : start
  **/   
  
-#include <term/values.h>
+#include <sh/shell.h>
 #include <str_t.h>  
 
 char* VAR_USER = "root";
